+++ component/async/sqs/option_test.go
