+++ examples/fifth/main.go
@@ -12,9 +12,9 @@ import (
 	"github.com/aws/aws-sdk-go/service/sqs"
 	"github.com/aws/aws-sdk-go/service/sqs/sqsiface"
 	"github.com/beatlabs/patron"
-	"github.com/beatlabs/patron/async"
-	patronsqs "github.com/beatlabs/patron/async/sqs"
 	patrongrpc "github.com/beatlabs/patron/client/grpc"
+	"github.com/beatlabs/patron/component/async"
+	patronsqs "github.com/beatlabs/patron/component/async/sqs"
 	"github.com/beatlabs/patron/component/grpc/greeter"
 	"github.com/beatlabs/patron/examples"
 	"github.com/beatlabs/patron/log"
