+++ examples/fourth/main.go
@@ -15,9 +15,9 @@ import (
 	"github.com/aws/aws-sdk-go/service/sqs"
 	"github.com/aws/aws-sdk-go/service/sqs/sqsiface"
 	"github.com/beatlabs/patron"
-	"github.com/beatlabs/patron/async"
-	"github.com/beatlabs/patron/async/amqp"
 	patronsns "github.com/beatlabs/patron/client/sns"
+	"github.com/beatlabs/patron/component/async"
+	"github.com/beatlabs/patron/component/async/amqp"
 	"github.com/beatlabs/patron/encoding/json"
 	"github.com/beatlabs/patron/examples"
 	"github.com/beatlabs/patron/log"
